requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
python-dotenv==1.0.0
langchain==0.0.335
openai==1.3.0
pypdf==3.17.1
python-docx==1.0.0
pytest==7.4.3
httpx==0.25.1 